requirements.txt
tqdm
opencv-python
gdown
Pillow
torch
torchvision
Lightning